% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqueous_humour.R
\docType{data}
\name{aqueous_humour}
\alias{aqueous_humour}
\title{Aqueous Humour data}
\format{
A data frame with 59 rows and 45 variables:
\describe{
  \item{ID}{ID observation}
  \item{group}{class membership (C, O)}
  \item{R1}{metabolic values}
  \item{R2}{metabolic values}
  \item{R3}{metabolic values}
  \item{R4}{metabolic values}
  \item{R5}{metabolic values}
  \item{R6}{metabolic values}
  \item{R7}{metabolic values}
  \item{R8}{metabolic values}
  \item{R9}{metabolic values}
  \item{R10}{metabolic values}
  \item{R11}{metabolic values}
  \item{R12}{metabolic values}
  \item{R13}{metabolic values}
  \item{R14}{metabolic values}
  \item{R15}{metabolic values}
  \item{R16}{metabolic values}
  \item{R17}{metabolic values}
  \item{R18}{metabolic values}
  \item{R19}{metabolic values}
  \item{R20}{metabolic values}
  \item{R21}{metabolic values}
  \item{R22}{metabolic values}
  \item{R23}{metabolic values}
  \item{R24}{metabolic values}
  \item{R25}{metabolic values}
  \item{R26}{metabolic values}
  \item{R27}{metabolic values}
  \item{R28}{metabolic values}
  \item{R29}{metabolic values}
  \item{R30}{metabolic values}
  \item{R31}{metabolic values}
  \item{R32}{metabolic values}
  \item{R33}{metabolic values}
  \item{R34}{metabolic values}
  \item{R35}{metabolic values}
  \item{R36}{metabolic values}
  \item{R37}{metabolic values}
  \item{R38}{metabolic values}
  \item{R39}{metabolic values}
  \item{R40}{metabolic values}
  \item{R41}{metabolic values}
  \item{R42}{metabolic values}
  \item{R43}{metabolic values}
}
}
\usage{
aqueous_humour
}
\description{
59 post-mortem aqueous humor samples collected from closed and opened sheep eyes
}
\references{
\url{https://link.springer.com/article/10.1007/s11306-019-1533-2}
}
\author{
Angela Andreella \email{angela.andreella@unive.it}
}
\keyword{datasets}
