% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSampleSize.R
\name{computeSampleSize}
\alias{computeSampleSize}
\title{sample size calculation}
\usage{
computeSampleSize(d, nperm, A, Nsim, post.transformation, scaling, seed, nperm, alpha = 0.05, eps = 0.01)
}
\arguments{
\item{d}{effect size value or list containing X and Y data}

\item{nperm}{number of permutations}

\item{A}{number of components}

\item{Nsim}{number of simulations}

\item{power}{power desired}

\item{N}{initial number of observations}

\item{post.transformation}{TRUE if you want to apply post transformation.}

\item{scaling}{type of scaling}

\item{seed}{seed}

\item{alpha}{level of significance}

\item{eps}{epsilon value}
}
\value{
Returns the corresponding pvalues
}
\description{
Compute optimal sample size calculation
}
\author{
Angela Andreella
}
