% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWT.R
\name{computeWT}
\alias{computeWT}
\title{Compute weight and score matrices from PLSc}
\usage{
computeWT(X, Y, A)
}
\arguments{
\item{X}{Data matrix where columns represent the \eqn{p} variables and
rows the \eqn{n} observations.}

\item{Y}{Data matrix where columns represent the two classes and
rows the \eqn{n} observations.}

\item{A}{Number of score components}
}
\value{
List with the following objects:
\describe{
  \item{W}{Matrix of weights}
  \item{T_score}{Matrix of \code{Y} scores}
  \item{R}{Matrix of \code{Y} residuals}
}
}
\description{
Compute weight and score matrices for Partial Least Squares classification
}
\author{
Angela Andreella
}
\keyword{internal}
