% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{crossValidation}
\alias{crossValidation}
\title{Optimal number of components by cross-validation}
\usage{
crossValidation(X, Y, K = 10, method = "cv", times = 1,seed = 123,
A = 5, randomization = FALSE, ...)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} variables and rows
the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} classes and rows
the \eqn{n} observations.}

\item{K}{number of folds}

\item{method}{\code{"cv"} or \code{"repeated-cv"}}

\item{times}{if \code{"method = repeated-cv"} here you specify the times}

\item{seed}{seed value}

\item{A}{number of maximum components}

\item{randomization}{if \code{TRUE} computes p-values. The harmonic mean
is considered across folds and times.}

\item{...}{Futher parameters.}
}
\value{
returns the mean performance across all folds and all repeats for
each components
}
\description{
Performs cross validation to find the optimal number of
components by MCC
}
\author{
Angela Andreella
}
