% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSize.R
\name{effectSize}
\alias{effectSize}
\title{effect size PLSc}
\usage{
effectSize(X, Y, A, measure, randomization, ...)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and
rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and
rows the \eqn{n} observations.}

\item{A}{number of components}

\item{measure}{now \code{"MCC"} and \code{"score"} are implemented.}

\item{randomization}{compute p-values?}

\item{...}{Futher parameters.}
}
\value{
Returns the corresponding pvalues
}
\description{
compute effect size based on MCC or predictive scores
}
\author{
Angela Andreella
}
