% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDA.R
\name{IDA}
\alias{IDA}
\title{Iteration Deflation Algorithm}
\usage{
IDA(X, Y, W)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{W}{weight matrix where columns represent the \eqn{A} components and rows the \eqn{k} X variables.}
}
\value{
Returns a matrix of scores vectors \code{Tscore}.
}
\description{
Performs Iteration Deflation Algorithm
}
\author{
Angela Andreella
}
