% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSc.R
\name{PLSc}
\alias{PLSc}
\title{PLS}
\usage{
PLSc(X, Y, A, scaling, post.transformation, eps, Y.prob, transformation = "clr")
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{A}{number of components}

\item{scaling}{type of scaling.}

\item{post.transformation}{TRUE if you want to apply post transformation.}

\item{eps}{parameter needed to transform dummy variable into a
vector of probabilities}

\item{Y.prob}{TRUE if Y describes the probability to being in the class.}

\item{transformation}{transformation between ilr and clr}
}
\value{
Returns a list with the following objects:
- \code{W}: matrix of weigths
- \code{X_loading}: matrix of X loading
- \code{Y_loading}: matirx of Y loading
- \code{X}: matrix of X data
- \code{Y}: matirx of Y data
- \code{T_score}: matrix of scores
- \code{Y_fitted}: fitted Y matrix
- \code{B}: Matrix regression coefficients
- \code{M}: number of orthogonal components if post transformation is applied.
}
\description{
Performs PLS two class
}
\author{
Angela Andreella
}
