% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Test.R
\name{R2Test}
\alias{R2Test}
\title{R2 test}
\usage{
R2Test(X, Y, nperm = 100, A, randomization = FALSE, Y.prob = FALSE, eps,...)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and
rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and
rows the \eqn{n} observations.}

\item{nperm}{number of permutations}

\item{A}{number of latent components}

\item{randomization}{compute p-values?}

\item{Y.prob}{Is Y a probability vector?}

\item{eps}{parameter needed to transform dummy variable into a}

\item{...}{Futher parameters.}
}
\value{
Returns a list with the corresponding statistical tests,
raw and adjusted p-values
}
\description{
Performs randomization test based on R2
}
\author{
Angela Andreella
}
