% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreTest.R
\name{scoreTest}
\alias{scoreTest}
\title{score test}
\usage{
scoreTest(X, Y, nperm, A, randomization, ...)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and
rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and
rows the \eqn{n} observations.}

\item{nperm}{number of permutations}

\item{A}{number of components}

\item{randomization}{compute p-values?}

\item{...}{Futher parameters.}
}
\value{
Returns a list with the corresponding statistical tests,
raw and adjusted p-values
}
\description{
Performs randomization test based on scores
}
\author{
Angela Andreella
}
