% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Tscore.R
\name{sim_Tscore}
\alias{sim_Tscore}
\title{simulate the T scores matrix}
\usage{
sim_Tscore(Tscore, n, seed)
}
\arguments{
\item{Tscore}{score matrix}

\item{n}{number of observations}

\item{seed}{fix seed}
}
\value{
Returns a simulated T score matrix.
}
\description{
simulate the T scores matrix from multivariate continuous
distribution by using kernel density estimation and the accept-reject method.
}
\author{
Angela Andreella
}
\keyword{internal}
