% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePilotData.R
\name{simulatePilotData}
\alias{simulatePilotData}
\title{simulate cluster data from normal distribution}
\usage{
simulatePilotData(seed=123, nvar, clus.size,m, s)
}
\arguments{
\item{seed}{seed value}

\item{nvar}{number of variables}

\item{clus.size}{size of classes (only two classes are considered)}

\item{m}{mean multivariate normal distribution}

\item{s}{sd multivariate normal distribution}
}
\value{
Returns list of X and Y simulated data
}
\description{
simulate simple cluster data from normal distribution
}
\author{
Angela Andreella
}
