% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umor_acqueo.R
\docType{data}
\name{umor_acqueo}
\alias{umor_acqueo}
\title{This is data to be included in my package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59 rows and 45 columns.
}
\usage{
umor_acqueo
}
\description{
This is data to be included in my package
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/31069551/}
}
\author{
Angela Andreella \email{angela.andreella@unive.it}
}
\keyword{datasets}
