% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheezing.R
\docType{data}
\name{wheezing}
\alias{wheezing}
\title{Wheezing data}
\format{
A data frame with 32 rows and 176 variables
}
\usage{
wheezing
}
\description{
32 urine samples from children at risk of early-onset asthma and those with transient wheezing.
}
\references{
\url{https://onlinelibrary.wiley.com/doi/10.1111/pai.12879}
}
\author{
Angela Andreella \email{angela.andreella@unive.it}
}
\keyword{datasets}
