% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePower.R
\name{computePower}
\alias{computePower}
\title{power computations}
\usage{
computePower(X, Y, A,
post.transformation = TRUE, scaling, n, seed = 123,
Nsim, nperm, alpha, eps = 0.01, test = "mcc", ...)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} variables and
rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} classes and
rows the \eqn{n} observations.}

\item{A}{number of latent components}

\item{post.transformation}{TRUE if you want to apply post transformation.}

\item{scaling}{type of scaling, one of
\code{c("auto-scaling", "pareto-scaling", "mean-centering")}}

\item{n}{sample size}

\item{seed}{seed value}

\item{Nsim}{number of simulations}

\item{nperm}{number of permutations}

\item{alpha}{type I error}

\item{eps}{parameter needed to transform dummy variable into a
vector of probabilities}

\item{test}{type of test, one of \code{c("score", "mcc", "eigen")}.
Default to "mcc".}

\item{...}{Futher parameters see \code{PLSc}}
}
\value{
Returns the corresponding estimated power
}
\description{
estimate power for a given sample size, alpha level and number
of latent components.
}
\author{
Angela Andreella
}
