% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePower.R
\name{computePower}
\alias{computePower}
\title{power computations}
\usage{
computePower(X, Y, A,
post.transformation, scaling, n, seed = 123,
Nsim, nperm, alpha, eps = 0.01, ...)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} variables and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} classes and rows the \eqn{n} observations.}

\item{A}{number of components}

\item{post.transformation}{TRUE if you want to apply post transformation.}

\item{scaling}{type of scaling.}

\item{n}{sample size}

\item{seed}{seed value}

\item{Nsim}{number of simulations}

\item{nperm}{number of permutations}

\item{alpha}{type I error}

\item{eps}{see details}

\item{test}{type of test}

\item{...}{Futher parameters.}
}
\value{
Returns the corresponding pvalues
}
\description{
Performs randomization test based on eigenvalues
}
\author{
Angela Andreella
}
