% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSampleSize.R
\name{computeSampleSize}
\alias{computeSampleSize}
\title{sample size calculation}
\usage{
computeSampleSize(n, X, Y, A, alpha, beta,
nperm, Nsim, seed, test = "R2",...)
}
\arguments{
\item{n}{vector of sample sizes to consider}

\item{X}{data matrix where columns represent the \eqn{p} variables and
rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the two classes and
rows the \eqn{n} observations.}

\item{A}{number of score components}

\item{alpha}{level of type I error. Default 0.05}

\item{beta}{level of type II error. Default 0.2.}

\item{nperm}{number of permutations. Default 100.}

\item{Nsim}{number of simulations. Default 100.}

\item{seed}{seed value}

\item{test}{type of test, one of \code{c("score", "mcc", "R2")}.
Default to @R2.}

\item{...}{Futher parameters.}
}
\value{
Returns a dataframe that contains the estimated power for each
sample size and number of components considered
}
\description{
Compute optimal sample size calculation
}
\examples{
\dontrun{
datas <- simulatePilotData(nvar = 30, clus.size = c(5,5),m = 6,nvar_rel = 5,ncomp = 2)
out <- computeSampleSize(X = datas$X, Y = datas$Y, A = 3, n = 20)
}
}
\author{
Angela Andreella
}
