% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSampleSize.R
\name{computeSampleSize}
\alias{computeSampleSize}
\title{sample size calculation}
\usage{
computeSampleSize(n, X, Y, A, post.transformation, alpha, beta,
nperm, Nsim, seed, scaling, ...)
}
\arguments{
\item{n}{two dimensional vector giving the minimum and maximum sample size
to be considered}

\item{X}{data matrix where columns represent the \eqn{p} variables and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} classes and rows the \eqn{n} observations.}

\item{A}{number of components}

\item{post.transformation}{TRUE if you want to apply post transformation.}

\item{alpha}{level of type I error}

\item{beta}{level of type II error}

\item{nperm}{number of permutations}

\item{Nsim}{number of simulations}

\item{seed}{seed value}

\item{scaling}{type of scaling.}

\item{...}{Futher parameters.}
}
\value{
Returns the corresponding pvalues
}
\description{
Compute optimal sample size calculation
}
\author{
Angela Andreella
}
