% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWT.R
\name{computeWT}
\alias{computeWT}
\title{compute weight and score matrix PLSc}
\usage{
computeWT(X, Y, A, scaling)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} variables and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{A}{number of components}
}
\value{
weight matrix \code{W}
}
\description{
compute weight matrix PLSc
}
\author{
Angela Andreella
}
