% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWT.R
\name{computeWT}
\alias{computeWT}
\title{compute weight and score matrices from PLSc}
\usage{
computeWT(X, Y, A)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} variables and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{A}{number of latent components}
}
\value{
Returns a list with the following objects:
- \code{W}: matrix of weigths
- \code{T_score}: matrix of Y scores
- \code{R}: matrix of Y residuals
}
\description{
compute weight and score matrices from PLSc
}
\author{
Angela Andreella
}
