% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenTest.R
\name{eigenTest}
\alias{eigenTest}
\title{randomization test}
\usage{
eigenTest(X, Y, nperm, A, scaling = "mean-centering", Y.prob = TRUE, eps = 0.01)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{nperm}{number of permutations}

\item{A}{number of components}

\item{scaling}{type of scaling}

\item{Y.prob}{TRUE if Y describes the probability to being in the class.}

\item{eps}{see details}
}
\value{
Returns the corresponding pvalues
}
\description{
Performs randomization test based on eigenvalues
}
\author{
Angela Andreella
}
