% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenTest.R
\name{eigenTest}
\alias{eigenTest}
\title{randomization test}
\usage{
eigenTest(X, Y, A, nperm, A, X.scaling, post.transformation)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{nperm}{number of permutations}

\item{A}{number of components}

\item{post.transformation}{TRUE if you want to apply post transformation.}
}
\value{
Returns the corresponding pvalues
}
\description{
Performs randomization test based on eigenvalues
}
\author{
Angela Andreella
}
