% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDA.R
\name{IDA}
\alias{IDA}
\title{Iteration Deflation Algorithm}
\usage{
IDA(X, Y, W)
}
\arguments{
\item{X}{Data matrix where columns represent the \eqn{p} variables and
rows the \eqn{n} observations.}

\item{Y}{Vector of class probabilities}

\item{W}{Weight matrix where columns represent the \eqn{A} components and rows the \code{X} variables. Computed from \code{\link{computeWT}}.}
}
\value{
Returns a matrix of scores vectors \code{Tscore}.
}
\description{
Performs Iteration Deflation Algorithm
}
\references{
Stocchero, M., & Paris, D. (2016). Post-transformation of PLS2 (ptPLS2) by orthogonal matrix: a new approach for generating predictive and orthogonal latent variables. Journal of Chemometrics, 30(5), 242-251.
}
\seealso{
\code{\link{PLSc}}, \code{\link{computeWT}}
}
\author{
Angela Andreella
}
\keyword{internal}
