% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSc.R
\name{PLSc}
\alias{PLSc}
\title{PLS}
\usage{
PLSc(X, Y, A, scaling, post.transformation = TRUE)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{A}{number of components}

\item{Y.prob}{TRUE if Y describes the probability to being in the class.}
}
\value{
Returns a list with the following objects:
- \code{W}: matrix of weigths
- \code{X.loading}: matrix of X loading
- \code{Y.loading}: matirx of Y loading
- \code{T_score}: matrix of scores
- \code{Y.fitted}: fitted Y matrix
- \code{B}: Matrix regression coefficients
- \code{M}: number of orthogonal components if post transformation is applied.
}
\description{
Performs PLS two class
}
\author{
Angela Andreella
}
