% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptPLSc.R
\name{ptPLSc}
\alias{ptPLSc}
\title{post transformed PLS}
\usage{
ptPLSc(X, Y, W)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} classes and rows the \eqn{n} observations.}

\item{Y}{data matrix where columns represent the \eqn{k} variables and rows the \eqn{n} observations.}

\item{W}{weight matrix where columns represent the \eqn{A} components and rows the \eqn{k} X variables.}
}
\value{
Returns a matrix of scores vectors \code{Tscore}.

Returns a list with the following objects:
- \code{W}: matrix of weigths
- \code{6}: post transformation matrix
- \code{M}: number of orthogonal components.
}
\description{
Performs post transformed PLS
}
\author{
Angela Andreella
}
