% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptPLSc.R
\name{ptPLSc}
\alias{ptPLSc}
\title{post transformed PLS}
\usage{
ptPLSc(X, Y, W)
}
\arguments{
\item{X}{data matrix where columns represent the \eqn{p} variables and
rows the \eqn{n} observations.}

\item{Y}{vector of class probabilities}

\item{W}{weight matrix where columns represent the \eqn{A} components and rows the \eqn{k} X variables.}
}
\value{
Returns a matrix of scores vectors \code{Tscore}.

Returns a list with the following objects:
- \code{W}: matrix of weights
- \code{6}: post transformation matrix
- \code{M}: number of orthogonal components.
}
\description{
Performs post transformed Partial Least Squares
}
\references{
Stocchero, M., & Paris, D. (2016). Post-transformation of PLS2 (ptPLS2) by orthogonal matrix: a new approach for generating predictive and orthogonal latent variables. Journal of Chemometrics, 30(5), 242-251.
}
\author{
Angela Andreella
}
\keyword{internal}
