% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_XY.R
\name{sim_XY}
\alias{sim_XY}
\title{simulate data matrix under the alternative hypothesis}
\usage{
sim_XY(out, n, seed = 123, post.transformation, A)
}
\arguments{
\item{out}{output from PLS2c}

\item{n}{number of observations, vector of length two (first values is the
number of observations for the first classe, the second for the second class)}

\item{seed}{fix seed}

\item{post.transformation}{TRUE if you want to apply post transformation.}

\item{A}{number of components}
}
\value{
Returns a simulated matrix under the alternative hypothesis.
}
\description{
simulate data matrix under the alternative hypothesis.
}
\author{
Angela Andreella
}
