% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePilotData.R
\name{simulatePilotData}
\alias{simulatePilotData}
\title{simulate cluster data from normal distribution}
\usage{
simulatePilotData(seed=123, nvar, clus.size, m, rho)
}
\arguments{
\item{seed}{seed value}

\item{nvar}{number of variables}

\item{clus.size}{size of classes (only two classes are considered)}

\item{m}{mean multivariate normal distribution}

\item{rho}{numeric value in `[0,1]`. Level of equi-correlation between pairs
of variables}
}
\value{
Returns list of X and Y simulated data
}
\description{
simulate simple cluster data from normal distribution
}
\author{
Angela Andreella
}
