% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePilotData.R
\name{simulatePilotData}
\alias{simulatePilotData}
\title{simulate pilot data}
\usage{
simulatePilotData(seed = 123, nvar, clus.size, nvar_rel,m, A = 2, S1 = NULL, S2 = NULL)
}
\arguments{
\item{seed}{seed value}

\item{nvar}{number of variables}

\item{clus.size}{Vector of two elements, specifying the size of classes (only two classes are considered)}

\item{nvar_rel}{number of variables relevant to predict \code{Y}}

\item{m}{separation between classes}

\item{A}{oracle number of score components}

\item{S1}{covariance matrix for the first class. Default @NULL i.e., the identity is considered.}

\item{S2}{covariance matrix for the second class. Default @NULL i.e., the identity is considered.}
}
\value{
Returns list of X and Y simulated data
}
\description{
simulate cluster pilot data
}
\examples{
datas <- simulatePilotData(nvar = 10, clus.size = c(5,5),m = 6,nvar_rel = 5,A = 2)
}
\author{
Angela Andreella
}
